Navigate with stack navigator to see restaurant info
