+++ package.json
@@ -11,7 +11,8 @@
     "axios": "^0.18.0",
     "react": "16.5.0",
     "react-native": "0.57.1",
-    "react-native-vector-icons": "^6.0.1"
+    "react-native-vector-icons": "^6.0.1",
+    "react-navigation": "^2.17.0"
   },
   "devDependencies": {
     "babel-jest": "23.6.0",
