+++ src/components/RestaurantInfo.js
@@ -1,0 +1,16 @@
+import React, { Component } from 'react'
+
+import {
+  View,
+  Text
+} from 'react-native'
+
+export default class RestaurantInfo extends Component {
+  render() {
+    return (
+      <View>
+        <Text>Info</Text>
+      </View>
+    )
+  }
+}
